CMSPlugin refactoring
Move application object management into a new ApplicationAwareTrait
